# openapi.json
{
  "components": {
    "schemas": {
...
      "ConfigApplyView": {
        "description": "`POST /config/apply` — apply-a-full-config result. The change is live but not written to disk.",
        "properties": {
          "applied": {
            "type": "boolean"
          },
          "config_version": {
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "applied",
          "config_version",
          "note"
        ],
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "enum": [
                  "not_found",
                  "unauthorized",
                  "method_not_allowed",
                  "forbidden",
                  "invalid_request",
                  "version_conflict",
                  "conflict",
                  "rate_limited",
                  "internal"
                ],
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "adminToken": {
        "in": "header",
        "name": "x-admin-token",
        "type": "apiKey"
      },
      "bearerAuth": {
        "description": "The same operator credential via Authorization: Bearer",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "The frozen, additive-only /api/v1/admin surface. Errors use the stable envelope {\"error\":{\"code\",\"message\"}}; tooling branches on `code`.",
    "title": "Busbar Admin API",
    "version": "1.5.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/config": {
      "get": {
        "operationId": "GetConfig",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EffectiveConfigView"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Missing/invalid admin credential (error code `unauthorized`)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Authenticated but under-scoped: requires `read-only` (error code `forbidden`)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal failure (error code `internal`); the detail is logged server-side, never returned"
          }
        },
        "security": [
          {
            "adminToken": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Effective running config snapshot (redacted)",
        "x-busbar-required-scope": "read-only"
      }
    },
    "/api/v1/admin/config/apply": {
      "post": {
        "operationId": "PostConfigApply",
        "parameters": [
          {
            "description": "Optimistic concurrency: the resource's ETag from a prior read (or the ETag returned by the previous mutation). Stale = 409 `version_conflict` (re-read and retry), nothing changes; absent or `*` = unconditional.",
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "description": "Replace the running configuration.",
                "properties": {
                  "config": {
                    "additionalProperties": true,
                    "description": "A `config.yaml` deploy block, as JSON. The accepted shape is the config file's own, documented in the configuration reference; it is not restated here because several of its types parse a wire shape that does not match their field layout.",
                    "type": "object"
                  },
                  "providers": {
                    "additionalProperties": true,
                    "description": "A `providers.yaml` document, as JSON. The accepted shape is the config file's own, documented in the configuration reference; it is not restated here because several of its types parse a wire shape that does not match their field layout.",
                    "type": "object"
                  }
                },
                "required": [
                  "config"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigApplyView"
                }
              }
            },
            "description": "`{applied, config_version, note}`"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "`invalid_request`: invalid config; nothing changed, malformed body / unknown field, malformed `If-Match` header"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Missing/invalid admin credential (error code `unauthorized`)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Authenticated but under-scoped: requires `full` (error code `forbidden`)"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "`version_conflict`: stale `If-Match` (re-read and retry)"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Per-principal mutation budget exhausted (error code `rate_limited`; `Retry-After` header)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal failure (error code `internal`); the detail is logged server-side, never returned"
          }
        },
        "security": [
          {
            "adminToken": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Apply a full config from the request body, atomically (live until next reload/restart; health preserved by lane identity)",
        "x-busbar-required-scope": "full"
      }
    },
...
    "/api/v1/admin/hooks": {
      "get": {
        "operationId": "GetHooks",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_HookView"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Missing/invalid admin credential (error code `unauthorized`)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Authenticated but under-scoped: requires `read-only` (error code `forbidden`)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal failure (error code `internal`); the detail is logged server-side, never returned"
          }
        },
        "security": [
          {
            "adminToken": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Hook registry (definitions)",
        "x-busbar-required-scope": "read-only"
      },
      "post": {
        "operationId": "PostHooks",
        "parameters": [
          {
            "description": "Optimistic concurrency: the resource's ETag from a prior read (or the ETag returned by the previous mutation). Stale = 409 `version_conflict` (re-read and retry), nothing changes; absent or `*` = unconditional.",
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "config": {
                    "additionalProperties": true,
                    "description": "A `hooks:` entry, as JSON. The accepted shape is the config file's own, documented in the configuration reference; it is not restated here because several of its types parse a wire shape that does not match their field layout.",
                    "type": "object"
                  },
                  "name": {
                    "description": "The hook name.",
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "config"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HookView"
                }
              }
            },
            "description": "Replaced — the name existed (same-grant re-register; body is the hook definition)"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HookView"
                }
              }
            },
            "description": "Registered — the name is NEW (body is the hook definition)"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "`invalid_request`: malformed body / unknown field, malformed `If-Match` header"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Missing/invalid admin credential (error code `unauthorized`)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "`forbidden`: a `hooks-register` principal may not touch a content-seeing (`prompt`/`user`) or `global` hook"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "`conflict`: base-defined (edit config.yaml), grant change on an existing definition | `version_conflict`: stale `If-Match` (re-read and retry)"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Per-principal mutation budget exhausted (error code `rate_limited`; `Retry-After` header)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal failure (error code `internal`); the detail is logged server-side, never returned"
          }
        },
        "security": [
          {
            "adminToken": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Register (or replace) a hook at runtime — live immediately",
        "x-busbar-required-scope": "hooks-register"
      }
    },
...
    "/api/v1/admin/info": {
      "get": {
        "operationId": "GetInfo",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfoView"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Missing/invalid admin credential (error code `unauthorized`)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Authenticated but under-scoped: requires `read-only` (error code `forbidden`)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal failure (error code `internal`); the detail is logged server-side, never returned"
          }
        },
        "security": [
          {
            "adminToken": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Version, compiled-in plugin proof, uptime, topology",
        "x-busbar-required-scope": "read-only"
      }
    },
...
  }
}